fastapi/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
click==8.1.5
dnspython==2.4.0
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.0
